requirements.txt
numpy
scipy
Pillow
ipython
pylint
pytest
matplotlib
opencv-python
torch
torchvision
scikit-learn
tqdm
sympy
pyautogui
pyobjc-core
pyobjc
